Restructuring and renaming of Pinnacle Export tool
