+++ packages/pymedphys/src/pymedphys/pinnacle.py
@@ -1,0 +1,2 @@
+from pymedphys_pinnacle import (
+    PinnacleExport)
